agent/requirements.txt
fastapi
uvicorn
torch
torchvision
torchxrayvision
pillow
cryptography